Factor common methods into base class.
