+++ src/samovar/generators/complete.py
@@ -36,44 +36,3 @@ class CompleteGenerator(BaseGenerator):
                     next_crop_of_states.append(new_state)
                     # TODO also record Event(rule, unifier) in there somewhere
                     # TODO if self.goal_is_met(new_state): goal_has_been_met = True (and indicate which one)
-
-    def generate_event(self):
-        candidates = self.get_candidate_rules()
-        if not candidates:
-            return None
-        rule, unifier = self.random.choice(candidates)
-        self.update_state(unifier, rule)
-        return Event(rule, unifier)
-
-    def get_candidate_rules(self, state):
-        candidates = []
-        for rule in self.scenario.rules:
-            for unifier in state.match_all(rule.pre.exprs, rule.pre.bindings):
-                candidates.append((rule, unifier))
-
-        if self.verbosity >= 3:
-            sys.stderr.write("Candidate rules:\n")
-            for rule, unifiers in candidates:
-                sys.stderr.write("-> " + rule.to_json() + "\n")
-                sys.stderr.write("---> {}\n".format(unifiers))
-            sys.stderr.write("")
-
-        return candidates
-
-    def update_state(self, state, env, rule):
-        new_state = state.clone()
-        for expr in rule.post.exprs:
-            term = expr.term.subst(env)
-            if isinstance(expr, Assert):
-                new_state.add(term)
-            elif isinstance(expr, Retract):
-                new_state.remove(term)
-        if self.verbosity >= 3:
-            self.debug_state("Intermediate")
-        return new_state
-
-    def debug_state(self, label):
-        sys.stderr.write(":::: {} State [\n".format(label))
-        for term in sorted(self.state.contents):
-            sys.stderr.write("::::   {}\n".format(term))
-        sys.stderr.write(":::: ]\n")
